Update API to handle camera frame + reorganize file + modify makefile to generate library .a

+++ include/mfis_api.h
@@ -1,5 +1,5 @@
 /**
- * \file mfis_api.c
+ * \file mfis_api.h
  * \brief Header for Communication API between A53 and R7 CPUs
  * \author esoftthings
  */
@@ -18,28 +18,35 @@
 * Public Structures
 ******************************************************************************************/
 /**
- * \struct Str_t
- * \brief Objet chaîne de caractères.
+ * \struct mfis_api_cam_buffers_info_t
+ * \brief Pointers to a camera triple buffers
  *
- * Str_t est un petit objet de gestion de chaînes de caractères. 
- * La chaîne se termine obligatoirement par un zéro de fin et l'objet 
- * connait la taille de chaîne contient !
  */
 typedef struct
 {
    uint32_t buffer_size;
-   uint32_t *ptr_buf1;
-   uint32_t *ptr_buf2;
-   uint32_t *ptr_buf3;
+   uint8_t *ptr_buf1;
+   uint8_t *ptr_buf2;
+   uint8_t *ptr_buf3;
 }
 mfis_api_cam_buffers_info_t;
 
+/**
+ * \struct mfis_api_cam_buffers_t
+ * \brief Pointers to all camera buffers
+ *
+ */
 typedef struct
 {
     mfis_api_cam_buffers_info_t cam[MFIS_API_MAX_CAMERA];
 }
 mfis_api_cam_buffers_t;
 
+/**
+ * \struct mfis_api_cam_last_buffers_id_t
+ * \brief ID of the cameras last updated frames
+ *
+ */
 typedef struct
 {
     uint32_t last_buffer_id[MFIS_API_MAX_CAMERA];
